+++ packages/widgetbook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **FIX**: Skip decoding non-ascii characters in URLs. ([#1218](https://github.com/widgetbook/widgetbook/pull/1218) - by [@shigomany](https://github.com/shigomany))
 - **FIX**: Encode all fields values to allow reserved characters (e.g. commas, colons and curly brackets). ([#1214](https://github.com/widgetbook/widgetbook/pull/1214))
+- **Feat**: Allow custom theming/personalization of widgetbook. ([#1225](https://github.com/widgetbook/widgetbook/pull/1225))
 
 ## 3.8.1